reduce test to only alpha indices
In the test section, if the parameter is set to "all" to compute all indices (alpha+beta+functional) BUT it seems to ask for large amount of memory or something like that and it fail. Locally, I asked for computing alpha indices in one hand, then beta, then functional, and it was ok one by one for each!