Forward medenv version to use latest copernicusmarine toolbox
The toolbox for accessing the copernicus data recently changed. Most of the previously available data access (https://help.marine.copernicus.eu/en/articles/8612591-switching-from-old-to-new-services) are deprecated.

Copernicusmarine is the latest toolbox to be used and is now available on conda-forge. `medenv` starting from 1.1.0 uses the copernicusmarine toolbox and now that copernicusmarine toolbox is released on conda, we can use it in medenv with the galaxy ecology platform.